➖ Remove:  업로드 플로우를 고려한 공통 헤더로 네이밍 변경
